touchups

+++ item-catalog/src/components/Item/Item.js
@@ -1,11 +1,9 @@
 import React from "react";
-import {Card, CardContent, Divider, Button, IconButton} from "@mui/material";
+import {Card, CardContent, Divider, IconButton} from "@mui/material";
 import { useState } from "react";
 import "./Item.css"
 import Delete from "@mui/icons-material/Delete";
-// import AddCircleIcon from '@mui/icons-material/AddCircle';
 import StarIcon from '@mui/icons-material/Star';
-import StarBorderIcon from '@mui/icons-material/StarBorder';
 import RemoveCircleIcon from '@mui/icons-material/RemoveCircle';
 import CloseIcon from '@mui/icons-material/Close';
 import AddCircleIcon from '@mui/icons-material/AddCircle';
@@ -19,7 +17,7 @@ const Item = ({itemName, itemDesc, itemPoster, itemPrice, itemRating, id, dbID, 
 
     const [open, setOpen] = useState(false);
     const auth=useAuth();
-    // console.log(recentUpdate)
+    const user = auth.user;
 
     const openDesc = () => {
         setOpen(true);
@@ -38,54 +36,46 @@ const Item = ({itemName, itemDesc, itemPoster, itemPrice, itemRating, id, dbID, 
         }
     }
 
-    const deleteItem = async() => {
+    const deleteItem = () => {
         let id = dbID;
-        // await fetch("http://localhost:4000/items/deleteItems/" + id, {
-        //     method: "DELETE",
-        //     "Cache-Control": "no-cache"
-        // })
+        let confirmDelete = window.confirm("Are you sure you want to delete " + itemName + "?")
 
-
-        axios.delete("http://localhost:4000/items/deleteItems/" + id).then((res) => {
-            console.log(res.data)
-            console.log('Student successfully updated')
-          }).catch((error) => {
-            console.log(error)
-          })
+        if (confirmDelete === true)
+        {
+            axios.delete("http://localhost:4000/items/deleteItems/" + id).then((res) => {
+                console.log(res.data)
+                }).catch((error) => {
+                console.log(error)
+                })
+        }
     }
 
-    const increaseRating = async() => {
+    const increaseRating = () => {
         let id = dbID;
         
         let user = auth.user
         let data = {user}
 
-        axios.put("http://localhost:4000/items/increaseRating/" + id, data).then((res)=>{console.log(res)}
-        ).catch((error)=>{console.error(error)})
-        // await fetch("http://localhost:4000/items/increaseRating/" + id, {
-        //     method: "PUT",
-        //     headers: {
-        //         "Content-Type": "application/json"
-        //     },
-        //     "Cache-Control": "no-cache"
-        // })
+        axios.put("http://localhost:4000/items/increaseRating/" + id, data).then((res)=>{console.log(res);
+        }
+        ).catch((error)=>{
+            const errorMessage = JSON.parse(error.request.response)
+            console.error(errorMessage.msg); 
+            alert(errorMessage.msg);})
     }
 
-    const decreaseRating = async() => {
+    const decreaseRating = () => {
         let id = dbID;
         let user = auth.user
 
         let data = {user}
 
-        axios.put("http://localhost:4000/items/decreaseRating/" + id, data).then((res)=>{console.log(res)}
-        ).catch((error)=>{console.error(error)})
-        // await fetch("http://localhost:4000/items/decreaseRating/" + id, {
-        //     method: "PUT",
-        //     headers: {
-        //         "Content-Type": "application/json"
-        //     },
-        //     "Cache-Control": "no-cache"
-        // })
+        axios.put("http://localhost:4000/items/decreaseRating/" + id, data).then((res)=>{console.log(res);
+        }
+        ).catch((error)=>{
+            const errorMessage = JSON.parse(error.request.response)
+            console.error(errorMessage.msg); 
+            alert(errorMessage.msg);})
     }
 
     return(<>
@@ -103,34 +93,23 @@ const Item = ({itemName, itemDesc, itemPoster, itemPrice, itemRating, id, dbID, 
             <Divider/>
             <>
             {open ? 
-                // <div key={id}>
                 <>
                     <CardContent key={id} sx={{paddingBottom: '1px'}}>
                         <span style={{fontSize: '20px', display: 'flex', textAlign: 'center', justifyContent: 'center'}}>{itemDesc}</span>
                     </CardContent>
                     <CardContent sx={{display: 'flex', alignItems: 'center', paddingTop: '1px', justifyContent: 'center'}}>
                         <Tooltip title="Close description"><IconButton onClick={closeDesc}><CloseIcon color="info" fontSize="large"></CloseIcon></IconButton></Tooltip>
-                        {/* <Button sx={{borderRadius: '25px', border: "1px solid black"}} variant="contained" color="success" onClick={closeDesc}>Hide Description</Button> */}
                     </CardContent>
-                {/* </div> */}
                 </>
                 : 
-                
-                // <div key={id}>
                 <>
                     <CardContent key={id} sx={{display: 'flex', alignItems: 'center', justifyContent: 'center', padding: '4px'}}>
                     <Tooltip title="Open description"><IconButton onClick={openDesc}><InfoIcon color="info" fontSize="large"></InfoIcon></IconButton></Tooltip>
                     </CardContent>
-                </>
-                    
-                // </div>
-                
+                </>         
                 }
             </>
-            {/* <Divider/>
-            <CardContent><span style={{display: 'flex', textAlign: 'center', justifyContent: 'center'}}>{itemQuantity} exist</span></CardContent> */}
             <Divider/>
-            <div>  
             <CardContent style={{display: 'flex', flexDirection: 'column', textAlign: 'center', justifyContent: 'center'}}>
                 <span>{checkRating()}</span>
                 <div>
@@ -139,30 +118,18 @@ const Item = ({itemName, itemDesc, itemPoster, itemPrice, itemRating, id, dbID, 
                 </div>
                
             </CardContent>
-
-            </div>
-            {/* <Divider/>
-            <CardContent style={{display: 'flex', justifyContent: 'center', textAlign: 'center'}}>
-                <span style={{display: 'flex', textAlign: 'center', justifyContent: 'center'}}>Last Updated: {Date(lastUpdate)}</span>
-            </CardContent> */}
-            {/* <>
-            { */}
-                {/* // itemRating < 4 ? */}
-                {/* <div key={id}>
-                <Divider></Divider>
-                    <CardContent sx={{display: 'flex', alignItems: 'center', justifyContent: 'center'}}>
-                        <IconButton onClick={deleteItem}><Delete color="error"></Delete></IconButton>
+            {
+                user === itemPoster? 
+                <>
+                    
+                    <Divider></Divider>
+                    <CardContent sx={{display: 'flex', paddingTop: '5px', paddingBottom: '10px!important', alignItems: 'center', justifyContent: 'center'}}>
+                        <Tooltip title="Delete item"><IconButton onClick={deleteItem} ><Delete color="error" size='large'></Delete></IconButton></Tooltip>
                     </CardContent>
-
-                </div> */}
-                {/* // : */}
+                </> : 
                 
-                {/* <div>
-
-                </div> */}
-
-            {/* }
-            </> */}
+                <></>
+            }
          </Card>
     </>)
